refactor: keep tax ID rules separate
